+++ b_aws_cdk_parallel/deploy_command.py
@@ -2,9 +2,10 @@ import json
 import subprocess
 from typing import Optional, Dict
 
-from b_aws_cdk_parallel.cdk_arguments import CdkArguments
 from b_continuous_subprocess.continuous_subprocess import ContinuousSubprocess
 
+from b_aws_cdk_parallel.aws_cdk_stack import AwsCdkStack
+from b_aws_cdk_parallel.cdk_arguments import CdkArguments
 from b_aws_cdk_parallel.color_print import cprint
 from b_aws_cdk_parallel.deployment_type import DeploymentType
 from b_aws_cdk_parallel.print_colors import PrintColors
@@ -13,7 +14,7 @@ from b_aws_cdk_parallel.print_colors import PrintColors
 class DeployCommand:
     def __init__(
             self,
-            stack: str,
+            stack: AwsCdkStack,
             type: DeploymentType,
             path: Optional[str] = None,
             env: Optional[Dict[str, str]] = None,
@@ -36,10 +37,10 @@ class DeployCommand:
         # when deploying, we can reuse already synthesized templates with
         # assets that are in cdk.out dir. More on deployments:
         # https://taimos.de/blog/deploying-your-cdk-app-to-different-stages-and-environments
-        app_stack = f'--app "cdk.out/" "{self.__stack}"'
+        app_stack = f'--app "cdk.out/" "{self.__stack.aws_cdk_name}"'
 
         if self.__deployment_type == DeploymentType.DEPLOY:
-            command = f'cdk deploy {app_stack}'
+            command = f'cdk deploy {app_stack} --exclusively'
         elif self.__deployment_type == DeploymentType.DESTROY:
             command = f'cdk destroy {app_stack} -f'
         else:
@@ -47,7 +48,7 @@ class DeployCommand:
 
         # We want to ensure that each stack deployment can have its own separate output.
         # Otherwise some clashes may start happening. Also, this approach is easier for debugging.
-        command += f' --output=./cdk_stacks/{self.__stack}'
+        command += f' --output=./cdk_stacks/{self.__stack.aws_cdk_name}'
         # We want to see beautiful continuous output, hence specify progress events.
         command += ' --progress events'
         # We do not want to interact with CLI, hence add "force" flag.
@@ -65,10 +66,10 @@ class DeployCommand:
 
         try:
             for line in process_generator:
-                print(f'[{self.__stack}]\t{line}', end='')
-            cprint(PrintColors.OKGREEN, f'Deployment of stack {self.__stack} was successful.')
+                print(f'[{str(self.__stack)}]\t{line}', end='')
+            cprint(PrintColors.OKGREEN, f'Deployment of stack {str(self.__stack)} was successful.')
         except subprocess.CalledProcessError as ex:
-            cprint(PrintColors.FAIL, f'Exception raised in {self.__stack} stack. Error: {repr(ex)}.')
+            cprint(PrintColors.FAIL, f'Exception raised in {str(self.__stack)} stack. Error: {repr(ex)}.')
 
             error_output = json.loads(ex.output)