+++ b_aws_cdk_parallel/stack_dependencies.py
@@ -1,8 +1,11 @@
 import json
 from typing import Optional, Dict, List, Any
 
+from b_aws_cdk_parallel.aws_cdk_stack import AwsCdkStack
 from b_aws_cdk_parallel.cdk_arguments import CdkArguments
 from b_aws_cdk_parallel.cdk_synth import CdkSynth
+from b_aws_cdk_parallel.color_print import multi_cprint
+from b_aws_cdk_parallel.print_colors import PrintColors
 
 
 class StackDependencies:
@@ -12,7 +15,7 @@ class StackDependencies:
             environment: Optional[Dict[str, str]] = None,
             reverse_dependencies: bool = False,
             cdk_arguments: Optional[CdkArguments] = None
-    ) -> Dict[str, List[str]]:
+    ) -> Dict[AwsCdkStack, List[AwsCdkStack]]:
         """
         Generates a dependency graph from a given AWS CDK application.
         It firstly synthesizes the application and then analyzes the
@@ -27,6 +30,9 @@ class StackDependencies:
 
         :return: Dependency graph.
         """
+        # If path is not set, assume "current path".
+        path = path or '.'
+
         CdkSynth.execute(path, environment, cdk_arguments)
 
         with open(f'{path}/cdk.out/manifest.json', 'r') as file:
@@ -41,7 +47,7 @@ class StackDependencies:
         return graph
 
     @staticmethod
-    def __stack_dependency_graph(cdk_manifest_json: Dict[str, Any]) -> Dict[str, List[str]]:
+    def __stack_dependency_graph(cdk_manifest_json: Dict[str, Any]) -> Dict[AwsCdkStack, List[AwsCdkStack]]:
         """
         Creates a stack dependency graph out of generated cdk manifest json file.
 
@@ -59,26 +65,35 @@ class StackDependencies:
                 "Stack4": []     # Stack4 has no dependencies i.e. Stack4 does not use other stacks.
             }
         """
-        # Determine which are stacks and which are not.
-        stacks = []
+        # Extract all stack resources from the cdk manifest json file.
+        # Create a lookup dictionary, where:
+        #   Keys - are aws cdk resource (stack) names.
+        #   Values - are aws cdk stack (`AwsCdkStack`) instances.
+        stacks_lookup: Dict[str, AwsCdkStack] = {}
         for name, artifact in cdk_manifest_json['artifacts'].items():
             if artifact['type'] == 'aws:cloudformation:stack':
-                stacks.append(name)
+                stacks_lookup[name] = AwsCdkStack(
+                    aws_cdk_name=name,
+                    pretty_name=artifact.get('properties', {}).get('stackName'),
+                    display_name=artifact.get('displayName')
+                )
 
         # Generate dependency graph.
-        stack_dependency_graph = {}
+        stack_dependency_graph: Dict[AwsCdkStack, List[AwsCdkStack]] = {}
         for name, artifact in cdk_manifest_json['artifacts'].items():
             if artifact['type'] == 'aws:cloudformation:stack':
                 # Get all of the dependencies - both stacks and other resources.
                 dependencies = artifact.get('dependencies', [])
                 # Filter just the stack dependencies because we don't care about other resources.
-                dependencies = [d for d in dependencies if d in stacks]
-                stack_dependency_graph[name] = dependencies
+                dependencies = [stacks_lookup[d] for d in dependencies if stacks_lookup.get(d)]
+                stack_dependency_graph[stacks_lookup[name]] = dependencies
 
         return stack_dependency_graph
 
     @staticmethod
-    def __reverse_dependency_graph(stack_dependency_graph: Dict[str, List[str]]) -> Dict[str, List[str]]:
+    def __reverse_dependency_graph(
+            stack_dependency_graph: Dict[AwsCdkStack, List[AwsCdkStack]]
+    ) -> Dict[AwsCdkStack, List[AwsCdkStack]]:
         """
         Creates a reverse stack dependency graph out of created stack dependency graph.
 
@@ -107,13 +122,16 @@ class StackDependencies:
         return reverse_stack_dependency_graph
 
     @staticmethod
-    def remove_dependency(stack: str, stack_dependency_graph: Dict[str, List[str]]) -> None:
+    def remove_dependency(
+            stack: AwsCdkStack,
+            stack_dependency_graph: Dict[AwsCdkStack, List[AwsCdkStack]]
+    ) -> None:
         """
         Removes a given stack from the dependency graph. This function does not
         return anything. It rather has a "side effect" that modifies the
         supplied stack dependency graph.
 
-        :param stack: Stack name to be removed.
+        :param stack: Stack to be removed.
         :param stack_dependency_graph: Graph from which the stack must be removed.
 
         :return: No return.
@@ -124,3 +142,18 @@ class StackDependencies:
         for _, dependencies in stack_dependency_graph.items():
             if stack in dependencies:
                 dependencies.remove(stack)
+
+    @staticmethod
+    def visualise_graph(stack_dependency_graph: Dict[AwsCdkStack, List[AwsCdkStack]]):
+        for stack, dependencies in stack_dependency_graph.items():
+            dependencies = ', '.join([str(d) for d in dependencies])
+            if len(dependencies) == 0:
+                multi_cprint(
+                    (PrintColors.OKGREEN, '»'),
+                    (PrintColors.OKBLUE, f' {stack}: [{dependencies}]')
+                )
+            else:
+                multi_cprint(
+                    (PrintColors.FAIL, '×'),
+                    (PrintColors.OKBLUE, f' {stack}: [{dependencies}]')
+                )